refactor(config): derive sandbox settings from schemas
